requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
deep-translator==1.11.4
idna==3.10
pyTelegramBotAPI==4.23.0
requests==2.32.3
schedule==1.2.2
soupsieve==2.6
telebot==0.0.5
urllib3==2.2.3
python-dotenv